"Added dynamic routing to service endpoints
This commit consists of the addition of dynamic routing methods to the application. This was done to enhance the flexibility of the app by handling requests with variable route parameters. Now, the app can redirect POST and GET requests to different services, ports, and any number of following parts, in a scalable and organized manner."