"Added selective debug print statement and reduced overall logging
A debug print statement has been added to selectively log response text when URLs with 'federate' are requested. This is to aid troubleshooting for these specific requests. Also, print statements have been removed from several other places in app.py to reduce the overall amount of logging and improve the readability of logs."