Refactor redirectGet3 method in app.py
The 'args' and 'body' variables in the 'redirectGet3' method were under-utilized or misused. The 'args' variable was assigned but never used resulting in wasted memory space. 'body' was set as an empty dictionary, however 'requests.get' doesn't send a body. Thus, 'body' variable is now set to None. This commit boosts efficiency and readability.